Combining two backports because the second has a test fix for the first.
SERVER-33078 convertToCapped size is not checked for float -> long long overflow

(cherry picked from commit f527188c4dab4bf2a3ce0e31406dbd121c3a90ca)

SERVER-35410 add incompatible_with_embedded tag to convert_to_capped.js

(cherry picked from commit 5377292873f5b203c401384567457b2987502528)

+++ src/mongo/db/catalog/capped_utils.cpp
@@ -113,7 +113,7 @@ mongo::Status mongo::cloneCollectionAsCapped(OperationContext* opCtx,
                                              Database* db,
                                              const std::string& shortFrom,
                                              const std::string& shortTo,
-                                             double size,
+                                             long long size,
                                              bool temp) {
     NamespaceString fromNss(db->name(), shortFrom);
     NamespaceString toNss(db->name(), shortTo);
@@ -251,7 +251,7 @@ mongo::Status mongo::cloneCollectionAsCapped(OperationContext* opCtx,
 
 mongo::Status mongo::convertToCapped(OperationContext* opCtx,
                                      const NamespaceString& collectionName,
-                                     double size) {
+                                     long long size) {
     StringData dbname = collectionName.db();
     StringData shortSource = collectionName.coll();
